Fix type of ContextMenu.CurrentEventId

+++ FFXIVClientStructs/FFXIV/Client/UI/Agent/AgentContext.cs
@@ -101,7 +101,7 @@ public unsafe partial struct AgentContext {
 [StructLayout(LayoutKind.Explicit, Size = 0x6A8)]
 public unsafe partial struct ContextMenu {
     [FieldOffset(0x00)] public short CurrentEventIndex;
-    [FieldOffset(0x02)] public short CurrentEventId;
+    [FieldOffset(0x02)] public byte CurrentEventId;
 
     [FieldOffset(0x08), FixedSizeArray] internal FixedSizeArray34<AtkValue> _eventParams; // 32 * AtkValue + 1 * AtkValue for submenus with title + 1 * unk
